Move out all helper functions to separate file
