+++ src/collections.ts
@@ -1,4 +1,5 @@
-import { addReturnSubmit, Engine } from "./engine";
+import { Engine } from "./engine";
+import { addReturnSubmit } from "./helpers";
 import { CollectionToolbar } from "./toolbars/collection-toolbar";
 import { AVL } from "./trees/AVL";
 import { BST } from "./trees/BST";
