+++ src/heaps/BinaryHeap.ts
@@ -1,4 +1,5 @@
-import { compare, Engine, MessagesObject } from "../../src/engine";
+import { Engine, MessagesObject } from "../../src/engine";
+import { compare } from "../../src/helpers";
 import { BinaryDir, BinaryNode } from "../../src/objects/binary-node";
 import { DSArray } from "../../src/objects/dsarray";
 import { TextCircle } from "../../src/objects/text-circle";
